EDNS: type mismatch on min() ssize_t / int

+++ lib/rfc2671.c
@@ -11,7 +11,7 @@ rfc2671RROptPack(char *buf, size_t sz, ssize_t edns_sz)
     // EDNS OPT record says only what our DNS buffer size is so far.
     snprintf(opt.name, RFC1035_MAXHOSTNAMESZ, ".");
     opt.type = RFC1035_TYPE_OPT;
-    opt._class = min(edns_sz, SQUID_UDP_SO_RCVBUF-1);
+    opt._class = min(edns_sz, (ssize_t)SQUID_UDP_SO_RCVBUF-1);
     opt.ttl = 0; // relevant?
     opt.rdata = NULL;
     opt.rdlength = 0;
